# central-sso/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSO Authentication API",
        "description": "Single Sign-On API for multi-tenant authentication",
        "contact": {
            "name": "SSO Admin",
            "email": "admin@sso.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get all permissions",
                "operationId": "a49af5d462a91e9996aa5342b6de4ddd",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "include_system",
                        "in": "query",
                        "description": "Include system permissions",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    },
                    {
                        "name": "include_roles",
                        "in": "query",
                        "description": "Include roles count",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PermissionResponseDTO"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Create a new permission",
                "operationId": "6ef531e33f8fca9c42d3e8fcaf30d5b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Create Reports"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "reports.create"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Can create new reports"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "reports"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permission created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Permission created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PermissionResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get a specific permission",
                "operationId": "16db55c19d22dca89994e416d7dd28be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PermissionResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Update a permission",
                "operationId": "c5094a50888fb3f3f1254fea3140241a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Create Reports"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "reports.create"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Can create new reports"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "reports"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Permission updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PermissionResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cannot update system permission"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Delete a permission",
                "operationId": "d0a14d107d6e379b9b3b5207f5c76e8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Permission deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cannot delete system permission"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permissions/categories": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get all permission categories",
                "operationId": "4ce1a133b412d635004855347a3fb188",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "users",
                                                "roles",
                                                "tenants",
                                                "system",
                                                "api"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PermissionResponseDTO": {
                "properties": {
                    "id": {
                        "description": "Permission ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Permission name",
                        "type": "string",
                        "example": "View Users"
                    },
                    "slug": {
                        "description": "Permission slug",
                        "type": "string",
                        "example": "users.view"
                    },
                    "description": {
                        "description": "Permission description",
                        "type": "string",
                        "example": "Allows viewing user list and details"
                    },
                    "category": {
                        "description": "Permission category",
                        "type": "string",
                        "example": "users"
                    },
                    "is_system": {
                        "description": "Whether the permission is a system permission",
                        "type": "boolean",
                        "example": false
                    },
                    "roles_count": {
                        "description": "Number of roles with this permission",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}